# src/openapi.yaml
openapi: 3.0.0
info:
  title: Problems Service
  version: 1.0.0
  description: The Problems Service is responsible for managing dsa and api sheets.
    DSA Problem Management - Create and manage traditional data structures and algorithms sheets with test cases and constraints.
    API Problem Management - Create API challenges with input/output contracts, expected implementations, and evaluation criteria.
    Category and Difficulty Tagging - Classify sheets based on difficulty levels, types, and categories for better discoverability.
servers:
  - url: /problems-svc/api/v1.0
    description: Local development server
components:
...
  schemas:
...
    registerSheetPayload:
      type: object
      properties:
        typeId:
          type: string
          description: Type ID
          minLength: 30
        title:
          type: string
          description: Title
          minLength: 1
        description:
          type: string
          description: Sheet Description
          minLength: 1
        difficulty:
          type: string
          minLength: 1
          enum:
            - Easy
            - Medium
            - Hard
        tags:
          type: array
          minItems: 1
          items:
            type: string
        examples:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/sheetExampleData'
        constraints:
          type: array
          minItems: 1
          items:
            type: string
        hints:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/sheetHintsData'
        testCases:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/testCasesData'
        codeSnippets:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/codeSnippetsData'
        referenceSolutions:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/referenceSolutionsData'
      required:
        - typeId
        - title
        - description
        - difficulty
        - tags
        - examples
        - constraints
        - hints
        - testCases
        - codeSnippets
        - referenceSolutions
    sheetExampleData:
      title: Schema for Sheet Example Data
      type: object
      properties:
        id:
          type: string
        input:
          type: string
          minLength: 1
        output:
          type: string
        explanation:
          type: string
          nullable: true
      required:
        - input
        - output
    sheetHintsData:
      title: Schema for Sheet Hints Data
      type: object
      properties:
        id:
          type: string
        orderNo:
          type: integer
        hint:
          type: string
          minLength: 1
      required:
        - orderNo
        - hint
    testCasesData:
      title: Schema for Sheet Test Cases
      type: object
      properties:
        input:
          type: string
          minLength: 1
        output:
          type: string
        isPublic:
          type: boolean
      required:
        - input
        - output
        - isPublic
    codeSnippetsData:
      title: Schema for Sheet Code Snippets
      type: object
      properties:
        id:
          type: string
        languageId:
          type: string
          minLength: 30
        snippet:
          type: string
          minLength: 1
      required:
         - languageId
         - snippet
    referenceSolutionsData:
      title: Schema for Sheet Reference Solutions
      type: object
      properties:
        languageId:
          type: string
          minLength: 30
        solution:
          type: string
          minLength: 1
      required:
        - languageId
        - solution
    sheetTagsData:
      title: Schema for Sheet Tags
      type: object
      properties:
        id:
          type: string
        tag:
          type: string
      required:
        - id
        - tag
    sheetTestCasesData:
      title: Schema for Sheet Test Cases
      type: object
      properties:
        id:
          type: string
        input:
          type: string
        output:
          type: string
        isPublic:
          type: boolean
      required:
        - id
        - input
        - output
...
    getSheetResponseData:
      title: Schema for retrieving Sheet Details
      type: object
      properties:
        id:
          type: string
        typeId:
          type: string
        sheetCode:
          type: integer
        title:
          type: string
        description:
          type: string
        difficulty:
          type: string
        constraints:
          type: array
          items:
            type: string
        createdDate:
          type: string
        modifiedDate:
          type: string
        tags:
          type: array
          items:
            $ref: '#/components/schemas/sheetTagsData'
        examples:
          type: array
          items:
            $ref: '#/components/schemas/sheetExampleData'
        hints:
          type: array
          items:
            $ref: '#/components/schemas/sheetHintsData'
        testCases:
          type: array
          items:
            $ref: '#/components/schemas/sheetTestCasesData'
        codeSnippet:
          type: array
          items:
            $ref: '#/components/schemas/codeSnippetsData'
        referenceSolutions:
          type: array
          items:
            $ref: '#/components/schemas/referenceSolutionsData'
      required:
        - id
        - sheetCode
        - title
        - description
        - difficulty
        - constraints
        - createdDate
        - modifiedDate
        - tags
        - examples
        - hints
        - testCases
        - codeSnippet
    getAllSheetsResponseData:
      title: Schema for retrieving all Sheet Details
      type: object
      properties:
        id:
          type: string
        typeId:
          type: string
        sheetCode:
          type: integer
        title:
          type: string
        difficulty:
          type: string
        tags:
          type: array
          items:
            $ref: '#/components/schemas/sheetTagsData'
      required:
        - id
        - typeId
        - sheetCode
        - title
        - difficulty
        - tags
...
    getSheetByIDResponse:
      type: object
      properties:
        statusCode:
          type: integer
          example: 201
        type:
          type: string
          example: 'SUCCESS'
        message:
          type: string
          example: 'Success'
        devMessage:
          type: string
          example: 'Success'
        success:
          type: boolean
          example: true
        data:
          type: object
          $ref: '#/components/schemas/getSheetResponseData'
      required:
        - statusCode
        - type
        - message
        - devMessage
        - success
        - data
    getAllSheetsResponse:
      type: object
      properties:
        statusCode:
          type: integer
          example: 201
        type:
          type: string
          example: 'SUCCESS'
        message:
          type: string
          example: 'Success'
        devMessage:
          type: string
          example: 'Success'
        success:
          type: boolean
          example: true
        data:
          type: array
          items:
            $ref: '#/components/schemas/getAllSheetsResponseData'
      required:
        - statusCode
        - type
        - message
        - devMessage
        - success
        - data
...
    noContentResponse:
      type: object
      properties:
        statusCode:
          type: integer
          example: 204
        type:
          type: string
          example: 'CONTENT_NOT_AVAILABLE'
        message:
          type: string
          example: 'No content available'
        devMessage:
          type: string
          example: 'No content available'
        success:
          type: boolean
          example: true
        data:
          type: array
    unauthorizedResponse:
      type: object
      properties:
        statusCode:
          type: integer
          example: 401
        type:
          type: string
          example: 'UNAUTHORIZED'
        message:
          type: string
          example: 'Unauthorized'
        devMessage:
          type: string
          example: Unauthorized access. Could not proceed with the request.
        error:
          type: array
        data:
          type: array
    badRequestResponse:
      type: object
      properties:
        statusCode:
          type: integer
          example: 400
        type:
          type: string
          example: 'BAD_REQUEST'
        message:
          type: string
          example: 'Bad Request'
        devMessage:
          type: string
          example: Bad Request. Required resource does not exist.
        error:
          type: array
        data:
          type: array
    forbiddenResponse:
      type: object
      properties:
        statusCode:
          type: integer
          example: 403
        type:
          type: string
          example: 'FORBIDDEN'
        message:
          type: string
          example: 'Forbidden'
        devMessage:
          type: string
          example: Forbidden access. Could not proceed with the request.
        error:
          type: array
        data:
          type: array
    notFound:
      type: object
      properties:
        statusCode:
          type: integer
          example: 404
        type:
          type: string
          example: 'NOT_FOUND'
        message:
          type: string
          example: 'Resource Not Found'
        devMessage:
          type: string
          example: Not Found. Requested resource does not exist.
        error:
          type: array
        data:
          type: array
    requestTimeout:
      type: object
      properties:
        statusCode:
          type: integer
          example: 408
        type:
          type: string
          example: 'REQUEST_TIMEOUT'
        message:
          type: string
          example: 'Request timeout.'
        devMessage:
          type: string
          example: 'Request timeout.'
        error:
          type: array
        data:
          type: array
    userAlreadyExist:
      type: object
      properties:
        statusCode:
          type: integer
          example: 409
        type:
          type: string
          example: 'CONFLICT'
        message:
          type: string
          example: 'Resource already exist'
        devMessage:
          type: string
          example: 'Resource already exist'
        error:
          type: array
        data:
          type: array
    unsupportedMedia:
      type: object
      properties:
        statusCode:
          type: integer
          example: 415
        type:
          type: string
          example: 'UNSUPPORTED_MEDIA'
        message:
          type: string
          example: 'The media format of the requested data is not supported by the server.'
        devMessage:
          type: string
          example: 'The media format of the requested data is not supported by the server.'
        error:
          type: array
        data:
          type: array
    unprocessableContent:
      type: object
      properties:
        statusCode:
          type: integer
          example: 422
        type:
          type: string
          example: 'UNPROCESSABLE_CONTENT'
        message:
          type: string
          example: 'The request was unabled to be followed due to semantic/internal errors.'
        devMessage:
          type: string
          example: 'The request was unabled to be followed due to semantic/internal errors.'
        error:
          type: array
        data:
          type: array
    internalServerErrorResponse:
      type: object
      properties:
        statusCode:
          type: integer
          example: 500
        type:
          type: string
          example: 'Internal Server Error'
        message:
          type: string
          example: 'Internal Server Error'
        devMessage:
          type: string
          example: 'Internal Server Error'
        error:
          type: array
        data:
          type: array
paths:
...
  /sheet:
    post:
      operationId: registerSheet
      summary: Register new Sheets
      description: An API to register new sheet in system
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/registerSheetPayload'
      responses:
        '201':
          description: REQUEST_COMPLETED
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/getSheetByIDResponse'
        '400':
          description: BAD_REQUEST
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/badRequestResponse'
        '401':
          description: UNAUTHORIZED
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/unauthorizedResponse'
        '403':
          description: FORBIDDEN
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/forbiddenResponse'
        '404':
          description: NOT_FOUND
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/notFound'
        '408':
          description: REQUEST_TIMEOUT
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/requestTimeout'
        '409':
          description: CONFLICT
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/userAlreadyExist'
        '415':
          description: UNSUPPORTED_MEDIA
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/unsupportedMedia'
        '422':
          description: UNPROCESSABLE_CONTENT
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/unprocessableContent'
        '500':
          description: INTERNAL_SERVER_ERROR
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internalServerErrorResponse'

    get:
      operationId: getAllSheets
      summary: Get all Sheets
      description: An API to retrieve all sheets basic info from system
      parameters:
        - in: query
          name: type
          schema:
            type: string
          required: false
          description: Filter sheets by sheet type
        - in: query
          name: tag
          schema:
            type: string
          required: false
          description: Filter sheets by tag
        - in: query
          name: page
          schema:
            type: string
          required: false
          description: Filter sheets by page number
        - in: query
          name: limit
          schema:
            type: string
          required: false
          description: Filter sheets by page limit
      responses:
        '200':
          description: SUCCESS
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/getAllSheetsResponse'
        '204':
          description: CONTENT_NOT_AVAILABLE
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/noContentResponse'
        '400':
          description: BAD_REQUEST
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/badRequestResponse'
        '401':
          description: UNAUTHORIZED
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/unauthorizedResponse'
        '403':
          description: FORBIDDEN
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/forbiddenResponse'
        '404':
          description: NOT_FOUND
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/notFound'
        '500':
          description: INTERNAL_SERVER_ERROR
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internalServerErrorResponse'
...